requirements.txt
# NervaOS Complete Dependencies
# All-in-one installation for full functionality

# ═══════════════════════════════════════════════════════════════════
# CORE SYSTEM
# ═══════════════════════════════════════════════════════════════════
dbus-next>=0.2.3          # Async DBus communication
PyGObject>=3.44.0         # GTK4 bindings for UI
psutil>=5.9.0             # System monitoring (CPU, RAM, processes)
keyring>=24.0.0           # Secure storage (Gnome Keyring)

# ═══════════════════════════════════════════════════════════════════
# AI PROVIDERS
# ═══════════════════════════════════════════════════════════════════
google-genai>=0.3.0       # Google Gemini SDK (primary AI)
httpx>=0.25.0             # Async HTTP for OpenAI/Claude/Custom APIs
anthropic>=0.7.0          # Anthropic Claude SDK (optional)
openai>=1.3.0             # OpenAI GPT SDK (optional)

# ═══════════════════════════════════════════════════════════════════
# VOICE CONTROL (Full Conversation)
# ═══════════════════════════════════════════════════════════════════
deepgram-sdk>=3.0.0       # Deepgram speech-to-text (production-grade)
PyAudio>=0.2.13           # Audio capture from microphone
pyttsx3>=2.90             # Text-to-speech (offline, fast)
pygame>=2.5.0             # Audio playback for voice (mixer)
websockets>=12.0          # Deepgram streaming WebSocket (voice)
# pvporcupine>=2.2.0      # Wake word detection (optional premium)

# ═══════════════════════════════════════════════════════════════════
# AUTOMATION & WORKFLOWS
# ═══════════════════════════════════════════════════════════════════
PyYAML>=6.0               # YAML workflow definitions
schedule>=1.1.0           # Job scheduling
python-crontab>=3.0.0     # Cron-like scheduling

# ═══════════════════════════════════════════════════════════════════
# DATA STORAGE
# ═══════════════════════════════════════════════════════════════════
aiosqlite>=0.19.0         # Async SQLite (history, patterns, operations)
sqlalchemy>=2.0.0         # ORM for license server database

# ═══════════════════════════════════════════════════════════════════
# WEB SEARCH
# ═══════════════════════════════════════════════════════════════════
ddgs>=9.0                 # DuckDuckGo / metasearch (web_search integration)

# ═══════════════════════════════════════════════════════════════════
# UTILITIES
# ═══════════════════════════════════════════════════════════════════
pyxdg>=0.28               # XDG base directory support
python-dotenv>=1.0.0      # .env file configuration
python-dateutil>=2.8.0    # Date/time utilities
pillow>=10.0.0            # Image processing (for future features)
requests>=2.31.0          # HTTP requests (fallback)

# ═══════════════════════════════════════════════════════════════════
# LICENSE SERVER (Optional - for commercial deployments)
# ═══════════════════════════════════════════════════════════════════
fastapi>=0.104.0          # License server API
uvicorn>=0.24.0           # ASGI server
pydantic>=2.0.0           # Data validation
python-jose>=3.3.0        # JWT tokens
passlib>=1.7.4            # Password hashing
python-multipart>=0.0.6   # Form data handling

# ═══════════════════════════════════════════════════════════════════
# DEVELOPMENT & TESTING (Optional)
# ═══════════════════════════════════════════════════════════════════
# black>=23.0.0           # Code formatter
# mypy>=1.0.0             # Type checking
# pytest>=7.4.0           # Testing framework
# pytest-asyncio>=0.21.0  # Async test support
# pytest-cov>=4.1.0       # Coverage reporting

# ═══════════════════════════════════════════════════════════════════
# NOTES
# ═══════════════════════════════════════════════════════════════════
# System dependencies (install via apt):
#   - libgtk-4-dev gir1.2-gtk-4.0 libadwaita-1-dev
#   - portaudio19-dev python3-pyaudio
#   - espeak espeak-ng espeak-data
#   - libwnck-3-dev gir1.2-wnck-3.0
#   - libnotify-bin gnome-keyring
#
# The install.sh script handles all system dependencies automatically!
#
# Total size: ~500MB including all dependencies
# RAM usage: ~150-300MB depending on features used
